configurator: fix button hover states
Previously, multiple buttons shared the same touch target for hover states.
That meant that multiple buttons were shown as hovered when only one was
actually hovered.
The problem was that multiple inputs and buttons shared the same label.

This patch removes all obsolete and error causing labels, and moves the
preview container out of the "Preview" label.

Signed-off-by: Florian Scherf <mail@florianscherf.de>