installer: fix menu handling
Previously, the menu was handled as setting. When the menu was set by the app
using `app.setting.PICOCSS_MENU` and "Reset" was clicked in the configurator,
the menu was reset to the default menu.

This patch moves the setting `PICOCSS_MENU` to its own scope, so it can not
be overridden by the configurator.

Signed-off-by: Florian Scherf <mail@florianscherf.de>