templates: base: fix content height
Previously, the content had a dynamic height. This resulted in the scrollbar
popping in and out, on page loads, and when a popup opened. Also the footer
jumped to the top when a page got loaded.

This patch sets a min-height of 100vh to the Lona container, to ensure that
the scrollbar is visible all the time, and the footer is at the bottom.

Signed-off-by: Florian Scherf <mail@florianscherf.de>