static: update npm install
Previously, none of the node dependencies were part of the git, but were
installed at packaging time. That worked for local development, but meant that
install via GitHub URL would result in a broken installation that contained no
static files.

This patch changes the Makefile, so node dependencies get downloaded and added
to the work tree, and adds an initial version of all static files.

Signed-off-by: Florian Scherf <mail@florianscherf.de>