templates: base: fix reconnect feature
Previously, some features, for example the auto-close feature of the top
navigation, were dependent on the auto-recconect feature. This was no real
dependency but an implementation detail of this template.

This patch resolves this dependency, by adding a variable called
`auto_reconnect` to the template, so any feature and hook can check it
individually.
This patch also removes the `Server is reconnecting` string when the server
disconnected, but the reconnect feature is not enabled.

Signed-off-by: Florian Scherf <mail@florianscherf.de>