templates: base: auto reconnect: fix double reconnects
Previously, when performing an auto reconnect, the client would reopen the
websocket connection and in the case of success reload the tab. This reload is
crucial to ensure a connect and a reconnect result in the same user experience,
but had the side effect of accessing the same view twice, which creates
problems when debugging or reading the server logs.

Lona upstream fixed this problem by adding the `create_window` option to
`LonaContext.reconnect()`.

    Lona upstream patch: 0452feeb6cf2 ("client: add support for reconnect without creating a window")

This patch fixes the double reconnect problem by just using the new option.

Signed-off-by: Florian Scherf <mail@florianscherf.de>