settings: switch from templating to global module state
Previously, Lonas templating mechanism was used to generate the configured
theme for the templates. This works fine if the only target for the theme is
Lona templates, but in the future lona-picocss should be able to support other
templating mechanisms (for example Django DTL) as well.

This patch moves all configuration into global module state to add support for
other templating mechanisms in the future.

Signed-off-by: Florian Scherf <mail@florianscherf.de>