add NavDivider for non-selectable menu entries
Adds a `NavDivider` dataclass that can be used in a `nav_items` list
like a `NavItem`, only taking `title` (and `icon`). It will be rendered
in the pull-down with a grey background, and is unclickable.

Styling can be changed with CSS: `li[role=divider] { ...; }`

As Jinja2 cannot test if a variable is an instance (beyond some built-ins)
the `divider` attribute (default: `True`) is tested. Dataclasses do not
have the option to make an attribute immutable, so this can be broken
by assigning to that attribute.