installer: make debug a global setting; remove obsolete middleware
This change was initially made because the inline middleware class within the
installer function caused crashes with the new live-reload feature of Lona
upstream. Having a global `PICOCSS_DEBUG` setting is a good idea for the future
anyways.

This patch also updates the settings view because for now this setting can only
be set via code.

Signed-off-by: Florian Scherf <mail@florianscherf.de>